Added json config setup logic Made util layer refactoring
