+++ VCheckSDK/VCheckSDK/util/GlobalUtils.swift
@@ -19,25 +19,62 @@ public final class GlobalUtils {
     }
 }
 
-extension UIImage {
-  convenience init?(named: String) {
-      self.init(named: named, in: InternalConstants.bundle, compatibleWith: nil)
-  }
-}
-
-extension Text {
-init(textKey: LocalizedStringKey) {
-    self.init(textKey, bundle: InternalConstants.bundle)
-  }
-}
-
 internal struct InternalConstants {
     private class EmptyClass {}
     static let bundle = Bundle(for: InternalConstants.EmptyClass.self)
 }
 
-extension String {
-    public var localized: String {
-        return NSLocalizedString(self, tableName: nil, bundle: InternalConstants.bundle, value: "", comment: "")
+var bundleKey: UInt8 = 0
+
+class AnyLanguageBundle: Bundle {
+
+    override func localizedString(forKey key: String,
+                              value: String?,
+                              table tableName: String?) -> String {
+
+        guard let path = objc_getAssociatedObject(self, &bundleKey) as? String,
+            let bundle = Bundle(path: path) else {
+            return super.localizedString(forKey: key, value: value, table: tableName)
+        }
+        return bundle.localizedString(forKey: key, value: value, table: tableName)
     }
+}
+
+extension Bundle {
+
+    class func setLanguage(_ language: String) {
+        defer {
+            object_setClass(InternalConstants.bundle, AnyLanguageBundle.self)
+        }
+        objc_setAssociatedObject(InternalConstants.bundle, &bundleKey,
+                                 InternalConstants.bundle.path(forResource: language, ofType: "lproj"),
+            .OBJC_ASSOCIATION_RETAIN_NONATOMIC)
+    }
+}
+
+extension UIApplication {
+
+  static var topWindow: UIWindow {
+    if #available(iOS 15.0, *) {
+      let scenes = UIApplication.shared.connectedScenes
+      let windowScene = scenes.first as? UIWindowScene
+      return windowScene!.windows.first!
+    } else {
+        return UIApplication.shared.windows.filter { $0.isKeyWindow }.first!
+    }
+  }
+}
+
+extension UIDevice {
+    static var isSimulator: Bool = {
+        #if targetEnvironment(simulator)
+        return true
+        #else
+        return false
+        #endif
+    }()
+}
+
+enum VCheckError: Error {
+    case initError(String)
 }
